services/processor/requirements.txt
pandas
requests
yfinance
flask
python-dotenv
aiohttp
asyncio
grpcio>=1.76.0
certifi
